Update Adafruit_DHT.c
Previously, i had the file setup in a dumb way.  The crontab would run the program each minute, and redirect the output to a file. This was causing an issue because it would run and the file would be opened, then if it took 30 seconds to read the temp, the tempest website could not access the file at that time so it thought it was disconnected.  Now I have it just opening the file once the temp is read, writing it quickly and close the file pointer. this issue should no longer occur. tested and it works great.