Update Feasability.tex
Work in progress. Added some content. Fixed a metric shit ton of spelling errors. Added a couple of picture figures, although I haven't put the pictures in the directory itself yet to be compiled with.  Going for a jog, then dinner, then I'll be back to write more. And add more diagrams. And find out wth an MSP430 is and why its a better alternative.